Move some common headers to host/common
... to continue to decouple and break the circular dependencies
between the main host server and the gl and vk backends.

Bug: b/537737772
Test: bazel test //host/...
Low-Coverage-Reason: REFACTOR_ONLY
