+++ host/Android.bp
@@ -135,7 +135,6 @@ cc_defaults {
         "channel_stream.cpp",
         "color_buffer.cpp",
         "frame_buffer.cpp",
-        "hwc2.cpp",
         "post_worker.cpp",
         "post_worker_gl.cpp",
         "read_buffer.cpp",
