+++ host/BUILD.bazel
@@ -67,15 +67,9 @@ cc_library(
     hdrs = [
         "buffer.h",
         "color_buffer.h",
-        "compositor.h",
         "frame_buffer.h",
-        "framework_formats.h",
-        "handle.h",
-        "hwc2.h",
         "post_commands.h",
         "post_worker.h",
-        "readback_worker.h",
-        "stale_ptr_registry.h",
         "vsync_thread.h",
     ],
     copts = GFXSTREAM_HOST_COPTS,
@@ -103,7 +97,6 @@ cc_library(
         "channel_stream.cpp",
         "color_buffer.cpp",
         "frame_buffer.cpp",
-        "hwc2.cpp",
         "post_worker.cpp",
         "post_worker_gl.cpp",
         "read_buffer.cpp",
@@ -128,16 +121,11 @@ cc_library(
         "buffer.h",
         "channel_stream.h",
         "color_buffer.h",
-        "compositor.h",
         "frame_buffer.h",
-        "framework_formats.h",
-        "handle.h",
-        "hwc2.h",
         "post_commands.h",
         "post_worker.h",
         "post_worker_gl.h",
         "read_buffer.h",
-        "readback_worker.h",
         "render_channel_impl.h",
         "render_control.h",
         "render_lib_impl.h",
@@ -146,7 +134,6 @@ cc_library(
         "render_window.h",
         "renderer_impl.h",
         "ring_stream.h",
-        "stale_ptr_registry.h",
         "sync_thread.h",
         "virtgpu_gfxstream_protocol.h",
         "virtio_gpu.h",
