+++ host/common/Android.bp
@@ -29,6 +29,7 @@ cc_library_static {
         "file_stream.cpp",
         "graphics_driver_lock.cpp",
         "guest_operations.cpp",
+        "hwc2.cpp",
         "mem_stream.cpp",
         "renderer_operations.cpp",
         "stream_utils.cpp",
