+++ host/common/BUILD.bazel
@@ -18,6 +18,7 @@ cc_library(
         "file_stream.cpp",
         "graphics_driver_lock.cpp",
         "guest_operations.cpp",
+        "hwc2.cpp",
         "mem_stream.cpp",
         "renderer_operations.cpp",
         "stream_utils.cpp",
