+++ host/common/hwc2.cpp
@@ -12,13 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "hwc2.h"
+#include "gfxstream/host/hwc2.h"
 
 namespace gfxstream {
 namespace host {
 
-std::unique_ptr<FlatComposeRequest> ToFlatComposeRequest(
-        const ComposeDevice* composeRequest) {
+std::unique_ptr<FlatComposeRequest> ToFlatComposeRequest(const ComposeDevice* composeRequest) {
     auto flatComposeRequest = std::make_unique<FlatComposeRequest>();
     flatComposeRequest->displayId = 0;
     flatComposeRequest->targetHandle = composeRequest->targetHandle;
@@ -28,8 +27,7 @@ std::unique_ptr<FlatComposeRequest> ToFlatComposeRequest(
     return flatComposeRequest;
 }
 
-std::unique_ptr<FlatComposeRequest> ToFlatComposeRequest(
-        const ComposeDevice_v2* composeRequest) {
+std::unique_ptr<FlatComposeRequest> ToFlatComposeRequest(const ComposeDevice_v2* composeRequest) {
     auto flatComposeRequest = std::make_unique<FlatComposeRequest>();
     flatComposeRequest->displayId = composeRequest->displayId;
     flatComposeRequest->targetHandle = composeRequest->targetHandle;
