+++ host/common/include/gfxstream/host/handle.h
