+++ host/frame_buffer.cpp
@@ -45,6 +45,7 @@
 #include "gfxstream/containers/Lookup.h"
 #include "gfxstream/host/display_operations.h"
 #include "gfxstream/host/guest_operations.h"
+#include "gfxstream/host/hwc2.h"
 #include "gfxstream/host/native_sub_window.h"
 #include "gfxstream/host/renderer_operations.h"
 #include "gfxstream/host/stream_utils.h"
@@ -55,7 +56,6 @@
 #include "gfxstream/synchronization/Lock.h"
 #include "gfxstream/system/System.h"
 #include "host/gl/context_helper.h"
-#include "hwc2.h"
 #include "render-utils/MediaNative.h"
 #include "render_thread_info.h"
 #include "sync_thread.h"
