+++ host/frame_buffer.h
@@ -31,16 +31,16 @@
 
 #include "buffer.h"
 #include "color_buffer.h"
-#include "framework_formats.h"
 #include "gfxstream/AsyncResult.h"
 #include "gfxstream/EventNotificationSupport.h"
 #include "gfxstream/host/color_buffer_interface.h"
 #include "gfxstream/host/external_object_manager.h"
+#include "gfxstream/host/framework_formats.h"
 #include "gfxstream/host/gfxstream_format.h"
 #include "gfxstream/host/gl_enums.h"
 #include "gfxstream/host/process_resources.h"
 #include "gfxstream/host/vk_enums.h"
-#include "handle.h"
+#include "gfxstream/host/handle.h"
 #include "post_commands.h"
 #include "render-utils/Renderer.h"
 #include "render-utils/render_api.h"
