+++ host/gl/color_buffer_gl.h
@@ -31,8 +31,8 @@
 #include "gfxstream/ManagedDescriptor.h"
 #include "gfxstream/host/features.h"
 #include "gfxstream/host/gfxstream_format.h"
-#include "handle.h"
-#include "hwc2.h"
+#include "gfxstream/host/hwc2.h"
+#include "gfxstream/host/handle.h"
 #include "pixel_read_formats.h"
 #include "render-utils/Renderer.h"
 #include "render-utils/stream.h"
