+++ host/gl/emulated_egl_window_surface.h
@@ -27,7 +27,7 @@
 #include "gfxstream/host/color_buffer_interface.h"
 #include "gl/color_buffer_gl.h"
 #include "gl/emulated_egl_context.h"
-#include "handle.h"
+#include "gfxstream/host/handle.h"
 
 namespace gfxstream {
 namespace host {
