+++ host/gl/emulation_gl.h
@@ -30,7 +30,6 @@
 #include "OpenGLESDispatch/GLESv2Dispatch.h"
 #include "buffer_gl.h"
 #include "color_buffer_gl.h"
-#include "compositor.h"
 #include "compositor_gl.h"
 #include "context_helper.h"
 #include "display_gl.h"
@@ -39,6 +38,7 @@
 #include "emulated_egl_fence_sync.h"
 #include "emulated_egl_image.h"
 #include "emulated_egl_window_surface.h"
+#include "gfxstream/host/compositor.h"
 #include "gfxstream/host/display.h"
 #include "gfxstream/host/display_surface.h"
 #include "gfxstream/host/features.h"
