+++ host/gl/stale_ptr_registry.h
