+++ host/vulkan/display_vk.h
@@ -28,9 +28,9 @@
 #include "debug_utils_helper.h"
 #include "display_surface_vk.h"
 #include "gfxstream/host/display.h"
+#include "gfxstream/host/hwc2.h"
 #include "gfxstream/synchronization/Lock.h"
 #include "goldfish_vk_dispatch.h"
-#include "host/hwc2.h"
 #include "swap_chain_state_vk.h"
 
 // The DisplayVk class holds the Vulkan and other states required to draw a
