+++ host/vulkan/vk_common_operations.h
@@ -35,12 +35,12 @@
 #include "gfxstream/host/RenderDoc.h"
 #include "gfxstream/host/external_object_manager.h"
 #include "gfxstream/host/features.h"
+#include "gfxstream/host/framework_formats.h"
 #include "gfxstream/host/gfxstream_format.h"
 #include "gfxstream/host/global_state.h"
 #include "gfxstream/host/vk_enums.h"
 #include "gfxstream/memory/UdmabufCreator.h"
 #include "goldfish_vk_private_defs.h"
-#include "host/framework_formats.h"
 #include "render-utils/Renderer.h"
 #include "vk_format_support.h"
 #include "vk_utils.h"
